+++ src/init.c
@@ -48,9 +48,7 @@ extern SEXP _prioritizr_rcpp_list_to_matrix_indices(SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_new_optimization_problem(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_optimization_problem_as_list(SEXP);
 extern SEXP _prioritizr_rcpp_predefined_optimization_problem(SEXP);
-extern SEXP _prioritizr_rcpp_set_optimization_problem_lb(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_set_optimization_problem_shuffled(SEXP);
-extern SEXP _prioritizr_rcpp_set_optimization_problem_ub(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_sp_to_polyset(SEXP, SEXP, SEXP);
 extern SEXP _prioritizr_rcpp_str_tree_to_sparse_matrix(SEXP);
 
@@ -95,9 +93,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_prioritizr_rcpp_new_optimization_problem",                          (DL_FUNC) &_prioritizr_rcpp_new_optimization_problem,                          3},
     {"_prioritizr_rcpp_optimization_problem_as_list",                      (DL_FUNC) &_prioritizr_rcpp_optimization_problem_as_list,                      1},
     {"_prioritizr_rcpp_predefined_optimization_problem",                   (DL_FUNC) &_prioritizr_rcpp_predefined_optimization_problem,                   1},
-    {"_prioritizr_rcpp_set_optimization_problem_lb",                       (DL_FUNC) &_prioritizr_rcpp_set_optimization_problem_lb,                       3},
     {"_prioritizr_rcpp_set_optimization_problem_shuffled",                 (DL_FUNC) &_prioritizr_rcpp_set_optimization_problem_shuffled,                 1},
-    {"_prioritizr_rcpp_set_optimization_problem_ub",                       (DL_FUNC) &_prioritizr_rcpp_set_optimization_problem_ub,                       3},
     {"_prioritizr_rcpp_sp_to_polyset",                                     (DL_FUNC) &_prioritizr_rcpp_sp_to_polyset,                                     3},
     {"_prioritizr_rcpp_str_tree_to_sparse_matrix",                         (DL_FUNC) &_prioritizr_rcpp_str_tree_to_sparse_matrix,                         1},
     {NULL, NULL, 0}